move other columns to column file
