+++ assets/src/components/backups/objectstores/ObjectStores.tsx
@@ -11,8 +11,6 @@ import isEmpty from 'lodash/isEmpty'
 
 import { useTheme } from 'styled-components'
 
-import { createColumnHelper } from '@tanstack/react-table'
-
 import { VirtualItem } from '@tanstack/react-virtual'
 
 import { useSetPageHeaderContent } from '../../cd/ContinuousDeployment'
@@ -21,21 +19,15 @@ import {
   OBJECT_STORES_REL_PATH,
 } from '../../../routes/backupRoutesConsts'
 import { FullHeightTableWrap } from '../../utils/layout/FullHeightTableWrap'
-import { ObjectStore, useObjectStoresQuery } from '../../../generated/graphql'
-import { Edge, extendConnection } from '../../../utils/graphql'
-import { ColWithIcon } from '../../utils/table/ColWithIcon'
+import { useObjectStoresQuery } from '../../../generated/graphql'
+import { extendConnection } from '../../../utils/graphql'
 
 import { GqlError } from '../../utils/Alert'
 
 import { useSlicePolling } from '../../utils/tableFetchHelpers'
 
 import CreateObjectStore from './CreateObjectStore'
-import {
-  ObjectStoreCloudIcon,
-  getObjectStoreCloud,
-  objectStoreCloudToDisplayName,
-} from './utils'
-import { ColActions } from './ObjectStoreColumns'
+import { ColActions, ColName, ColProvider } from './ObjectStoreColumns'
 
 const POLL_INTERVAL = 10 * 1000
 
@@ -55,37 +47,7 @@ const BACKUPS_OBJECT_STORES_BASE_CRUMBS: Breadcrumb[] = [
   },
 ]
 
-const columnHelper = createColumnHelper<Edge<ObjectStore>>()
-
-export const columns = [
-  columnHelper.accessor(({ node }) => node, {
-    id: 'provider',
-    header: 'Provider',
-    meta: { gridTemplate: `240px` },
-    cell: ({ getValue }) => {
-      const cloud = getObjectStoreCloud(getValue())
-
-      if (!cloud) return null
-
-      return (
-        <ColWithIcon
-          truncateLeft
-          icon={<ObjectStoreCloudIcon cloud={cloud} />}
-        >
-          {objectStoreCloudToDisplayName[cloud]}
-        </ColWithIcon>
-      )
-    },
-  }),
-  columnHelper.accessor(({ node }) => node?.name, {
-    id: 'name',
-    header: 'Storage name',
-    enableSorting: true,
-    enableGlobalFilter: true,
-    cell: ({ getValue }) => getValue(),
-  }),
-  ColActions,
-]
+export const columns = [ColProvider, ColName, ColActions]
 
 export default function ObjectStores() {
   const theme = useTheme()
